Updated Burger

+++ Burger.cs
@@ -9,9 +9,9 @@ namespace FastFood
     class Burger
     {
         public string Ingredients { get; set; }
-        public int Price { get; set; }
+        public double Price { get; set; }
 
-        public Burger(string ingredients, int price)
+        public Burger(string ingredients, double price)
         {
             Ingredients = ingredients;
             Price = price;
